Intermediate checkin, code to be tested
In the process of moving the settings to EEPROM

+++ firmware/pic18f/likalibs.X/cosm_socket.h
@@ -11,7 +11,7 @@
 #ifndef _COSM_SOCKET_H_
 #define _COSM_SOCKET_H_
 
-void cosm_init(char* auth_key, unsigned int feed);
+void cosm_init(char* auth_key, unsigned long feed);
 void cosm_process_uart(char data);
 void cosm_report(char* name, signed short value, char scale);
